Add a RESET button to return to the initial GUI state.
The RESET button will take the game back to the condition before the start button was pressed. It must
test if the Play Gain button (re-start) is displayed, and set it to display:none if it is. The RESET is
not at all necessary for game functionality, but, it feels more complete to include it.